+++ src/private/quality/GetNovaPesterTestResultPath.ps1
@@ -1,0 +1,9 @@
+function Get-NovaPesterTestResultPath {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)][string]$ProjectRoot
+    )
+
+    return [System.IO.Path]::Join($ProjectRoot, 'artifacts', 'TestResults.xml')
+}
+