+++ src/private/update/GetNovaAvailableSemanticVersion.ps1
@@ -1,0 +1,13 @@
+function Get-NovaAvailableSemanticVersion {
+    [CmdletBinding()]
+    param(
+        [object]$VersionInfo
+    )
+
+    if ($null -eq $VersionInfo) {
+        return $null
+    }
+
+    return [semver]$VersionInfo.Version
+}
+