+++ src/private/update/GetNovaModuleUpdateLookupScript.ps1
@@ -1,0 +1,9 @@
+function Get-NovaModuleUpdateLookupScript {
+    [CmdletBinding()]
+    param()
+
+    $lookupScriptPath = Get-ResourceFilePath -FileName 'update/ModuleUpdateLookup.ps1.txt'
+    return Get-Content -LiteralPath $lookupScriptPath -Raw -ErrorAction Stop
+}
+
+