+++ src/private/update/TestNovaStableUpdateAvailable.ps1
@@ -1,0 +1,10 @@
+function Test-NovaStableUpdateAvailable {
+    [CmdletBinding()]
+    param(
+        [semver]$StableVersion,
+        [Parameter(Mandatory)][semver]$InstalledVersion
+    )
+
+    return $null -ne $StableVersion -and $StableVersion -gt $InstalledVersion
+}
+